+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Added support for Humidity sensor module
   - Updated Google Play Services to 7.8.0
   - Added support for Android Studio 1.3.1
+  - Added support for Android 6.0 SDK (Marshmallow)
   - Updated gradle to 1.3.0
 
 ### 0.2.0