+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Added support for Android 6.0 SDK (Marshmallow)
 - Updated gradle to 1.3.0
 
-### 0.2.0
+### 0.2.0 (July 30, 2015)
 - Added support for Bluetooth sensor module
 - DataInterface method getDataInString() has been renamed to getDataInCSV()
 - SKDataInterface has been renamed to SKSensorData
@@ -17,5 +17,5 @@
 - Added support for Android Studio 1.3
 - Updated Build Tools version to 22.0.1
 
-### 0.1.0
+### 0.1.0 (July 13, 2015)
 - Initial Release